refactor(tests): share helpers via tests/common, add VecTransport::shuffle_to
The 2-line shuffle() that ferries bytes from one VecTransport's
outgoing to another's incoming was duplicated in tests/loopback.rs and
examples/loopback_poll.rs. Promote it to a method on VecTransport so
both can call self.shuffle_to(other) without redefining the helper.

Move the AlwaysAck PdHandler and the toy Lcg RNG (both used by chaos
tests) into tests/common/mod.rs following Cargo's standard pattern.
chaos_bus.rs now does mod common; use common::{AlwaysAck, Lcg};.
